# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/financial_statements:
    get:
      tags:
        - FinancialStatements
      summary: List FinancialStatements
      description: Get a paginated list of FinancialStatements with optional filtering and sorting
      operationId: listFinancialStatements
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialStatementList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - FinancialStatements
      summary: Create FinancialStatement
      description: Create a new FinancialStatement
      operationId: createFinancialStatement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFinancialStatementRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialStatement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/full_reconciles:
    get:
      tags:
        - FullReconciles
      summary: List FullReconciles
      description: Get a paginated list of FullReconciles with optional filtering and sorting
      operationId: listFullReconciles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullReconcileList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - FullReconciles
      summary: Create FullReconcile
      description: Create a new FullReconcile
      operationId: createFullReconcile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFullReconcileRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullReconcile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    FinancialStatement:
      type: object
      required:
        - id
        - company_id
        - statement_number
        - statement_type
        - name
        - fiscal_year
        - as_of_date
        - is_comparative
        - status
        - balance_check
        - trial_balance_check
        - line_items
        - comparative_data
        - variance_data
        - notes
        - generation_parameters
        - currency
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        statement_number:
          type: string
          maxLength: 30
        statement_type:
          type: string
        name:
          type: string
          maxLength: 100
        fiscal_period_id:
          nullable: true
          type: string
          format: uuid
        fiscal_year:
          type: integer
          format: int32
        fiscal_month:
          nullable: true
          type: integer
          format: int32
        as_of_date:
          type: string
          format: date
        period_start:
          nullable: true
          type: string
          format: date
        period_end:
          nullable: true
          type: string
          format: date
        is_comparative:
          type: boolean
        comparative_period_id:
          nullable: true
          type: string
          format: uuid
        comparative_as_of_date:
          nullable: true
          type: string
          format: date
        status:
          type: string
        total_current_assets:
          nullable: true
          type: string
        total_non_current_assets:
          nullable: true
          type: string
        total_assets:
          nullable: true
          type: string
        total_current_liabilities:
          nullable: true
          type: string
        total_non_current_liabilities:
          nullable: true
          type: string
        total_liabilities:
          nullable: true
          type: string
        total_equity:
          nullable: true
          type: string
        balance_check:
          type: boolean
        balance_difference:
          nullable: true
          type: string
        total_revenue:
          nullable: true
          type: string
        total_cogs:
          nullable: true
          type: string
        gross_profit:
          nullable: true
          type: string
        gross_profit_margin:
          nullable: true
          type: string
        total_operating_expenses:
          nullable: true
          type: string
        operating_income:
          nullable: true
          type: string
        operating_margin:
          nullable: true
          type: string
        total_other_income:
          nullable: true
          type: string
        total_other_expenses:
          nullable: true
          type: string
        income_before_tax:
          nullable: true
          type: string
        tax_expense:
          nullable: true
          type: string
        net_income:
          nullable: true
          type: string
        net_profit_margin:
          nullable: true
          type: string
        cash_from_operations:
          nullable: true
          type: string
        cash_from_investing:
          nullable: true
          type: string
        cash_from_financing:
          nullable: true
          type: string
        net_cash_change:
          nullable: true
          type: string
        beginning_cash:
          nullable: true
          type: string
        ending_cash:
          nullable: true
          type: string
        total_debits:
          nullable: true
          type: string
        total_credits:
          nullable: true
          type: string
        trial_balance_check:
          type: boolean
        line_items:
          type: object
        comparative_data:
          type: object
        variance_data:
          type: object
        notes:
          type: object
        management_notes:
          nullable: true
          type: string
          maxLength: 2000
        generated_at:
          nullable: true
          type: string
          format: date-time
        generated_by:
          nullable: true
          type: string
          format: uuid
        generation_parameters:
          type: object
        reviewed_at:
          nullable: true
          type: string
          format: date-time
        reviewed_by:
          nullable: true
          type: string
          format: uuid
        approved_at:
          nullable: true
          type: string
          format: date-time
        approved_by:
          nullable: true
          type: string
          format: uuid
        published_at:
          nullable: true
          type: string
          format: date-time
        pdf_url:
          nullable: true
          type: string
          maxLength: 500
        excel_url:
          nullable: true
          type: string
          maxLength: 500
        currency:
          type: string
        metadata:
          type: object
...
    CreateFinancialStatementRequest:
      type: object
      required:
        - company_id
        - statement_number
        - statement_type
        - name
        - fiscal_year
        - as_of_date
        - is_comparative
        - status
        - balance_check
        - trial_balance_check
        - line_items
        - comparative_data
        - variance_data
        - notes
        - generation_parameters
        - currency
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        statement_number:
          type: string
          maxLength: 30
        statement_type:
          type: string
        name:
          type: string
          maxLength: 100
        fiscal_period_id:
          nullable: true
          type: string
          format: uuid
        fiscal_year:
          type: integer
          format: int32
        fiscal_month:
          nullable: true
          type: integer
          format: int32
        as_of_date:
          type: string
          format: date
        period_start:
          nullable: true
          type: string
          format: date
        period_end:
          nullable: true
          type: string
          format: date
        is_comparative:
          type: boolean
        comparative_period_id:
          nullable: true
          type: string
          format: uuid
        comparative_as_of_date:
          nullable: true
          type: string
          format: date
        status:
          type: string
        total_current_assets:
          nullable: true
          type: string
        total_non_current_assets:
          nullable: true
          type: string
        total_assets:
          nullable: true
          type: string
        total_current_liabilities:
          nullable: true
          type: string
        total_non_current_liabilities:
          nullable: true
          type: string
        total_liabilities:
          nullable: true
          type: string
        total_equity:
          nullable: true
          type: string
        balance_check:
          type: boolean
        balance_difference:
          nullable: true
          type: string
        total_revenue:
          nullable: true
          type: string
        total_cogs:
          nullable: true
          type: string
        gross_profit:
          nullable: true
          type: string
        gross_profit_margin:
          nullable: true
          type: string
        total_operating_expenses:
          nullable: true
          type: string
        operating_income:
          nullable: true
          type: string
        operating_margin:
          nullable: true
          type: string
        total_other_income:
          nullable: true
          type: string
        total_other_expenses:
          nullable: true
          type: string
        income_before_tax:
          nullable: true
          type: string
        tax_expense:
          nullable: true
          type: string
        net_income:
          nullable: true
          type: string
        net_profit_margin:
          nullable: true
          type: string
        cash_from_operations:
          nullable: true
          type: string
        cash_from_investing:
          nullable: true
          type: string
        cash_from_financing:
          nullable: true
          type: string
        net_cash_change:
          nullable: true
          type: string
        beginning_cash:
          nullable: true
          type: string
        ending_cash:
          nullable: true
          type: string
        total_debits:
          nullable: true
          type: string
        total_credits:
          nullable: true
          type: string
        trial_balance_check:
          type: boolean
        line_items:
          type: object
        comparative_data:
          type: object
        variance_data:
          type: object
        notes:
          type: object
        management_notes:
          nullable: true
          type: string
          maxLength: 2000
        generated_at:
          nullable: true
          type: string
          format: date-time
        generated_by:
          nullable: true
          type: string
          format: uuid
        generation_parameters:
          type: object
        reviewed_at:
          nullable: true
          type: string
          format: date-time
        reviewed_by:
          nullable: true
          type: string
          format: uuid
        approved_at:
          nullable: true
          type: string
          format: date-time
        approved_by:
          nullable: true
          type: string
          format: uuid
        published_at:
          nullable: true
          type: string
          format: date-time
        pdf_url:
          nullable: true
          type: string
          maxLength: 500
        excel_url:
          nullable: true
          type: string
          maxLength: 500
        currency:
          type: string
        metadata:
          type: object
...
    FinancialStatementList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FinancialStatement'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    FullReconcile:
      type: object
      required:
        - id
        - company_id
        - exchange_total
        - reconciled_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        exchange_total:
          type: string
        reconciled_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          nullable: true
          type: object
...
    CreateFullReconcileRequest:
      type: object
      required:
        - company_id
        - exchange_total
        - reconciled_at
      properties:
        company_id:
          type: string
          format: uuid
        exchange_total:
          type: string
        reconciled_at:
          type: string
          format: date-time
        metadata:
          nullable: true
          type: object
...
    FullReconcileList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FullReconcile'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []